refactor: split file manager listing routes
